+++ src/Buckle/Buckle/CodeAnalysis/Compilation.cs
@@ -2,18 +2,13 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using Buckle.CodeAnalysis.Binding;
-using Buckle.CodeAnalysis.Display;
 using Buckle.CodeAnalysis.Emitting;
 using Buckle.CodeAnalysis.Evaluating;
 using Buckle.CodeAnalysis.Symbols;
 using Buckle.CodeAnalysis.Syntax;
 using Buckle.Diagnostics;
-using Diagnostics;
-using static Buckle.CodeAnalysis.Display.DisplayTextSegment;
-using static Buckle.Utilities.MethodUtilities;
 
 namespace Buckle.CodeAnalysis;
 
@@ -179,119 +174,6 @@ public sealed class Compilation {
         return result;
     }
 
-    // TODO Consider moving these methods to an extensions class under the Display namespace
-    /// <summary>
-    /// Emits the parse tree of the compilation.
-    /// </summary>
-    /// <param name="text">Out.</param>
-    internal void EmitTree(DisplayText text) {
-        if (globalScope.mainMethod != null) {
-            EmitTree(globalScope.mainMethod, text);
-        } else if (globalScope.scriptMethod != null) {
-            EmitTree(globalScope.scriptMethod, text);
-        } else {
-            var program = GetProgram();
-
-            foreach (var pair in program.methodBodies.OrderBy(p => p.Key.name))
-                EmitTree(pair.Key, text);
-        }
-    }
-
-    /// <summary>
-    /// Emits the parse tree of a single <see cref="MethodSymbol" /> after attempting to find it based on name.
-    /// Note: this only searches for methods, so if the name of another type of <see cref="Symbol" /> is passed it
-    /// will not be found.
-    /// </summary>
-    /// <param name="name">
-    /// The name of the <see cref="MethodSymbol" /> to search for and then print. If not found, throws.
-    /// </param>
-    /// <param name="text">Out.</param>
-    internal void EmitTree(string name, DisplayText text) {
-        var program = GetProgram();
-        var pair = LookupMethodFromParentsFromName(program, name);
-        SymbolDisplay.DisplaySymbol(text, pair.Item1);
-        text.Write(CreateSpace());
-        DisplayText.DisplayNode(text, pair.Item2);
-    }
-
-    /// <summary>
-    /// Emits the parse tree of a single <see cref="Symbol" />.
-    /// </summary>
-    /// <param name="symbol"><see cref="Symbol" /> to be the root of the <see cref="SyntaxTree" /> displayed.</param>
-    /// <param name="text">Out.</param>
-    internal void EmitTree(Symbol symbol, DisplayText text) {
-        var program = GetProgram();
-
-        void WriteTypeMembers(ITypeSymbolWithMembers type, bool writeEnding = true) {
-            try {
-                ImmutableList<Symbol> members = ImmutableList<Symbol>.Empty;
-
-                if (type is StructSymbol s)
-                    members = program.structMembers[s];
-                else if (type is ClassSymbol c)
-                    members = program.classMembers[c];
-
-                text.Write(CreateSpace());
-                text.Write(CreatePunctuation(SyntaxKind.OpenBraceToken));
-                text.Write(CreateLine());
-                text.indent++;
-
-                foreach (var field in members.OfType<FieldSymbol>()) {
-                    SymbolDisplay.DisplaySymbol(text, field);
-                    text.Write(CreateLine());
-                }
-
-                text.Write(CreateLine());
-
-                foreach (var method in members.OfType<MethodSymbol>()) {
-                    SymbolDisplay.DisplaySymbol(text, method);
-                    text.Write(CreateLine());
-                }
-
-                text.Write(CreateLine());
-
-                foreach (var typeMember in members.OfType<TypeSymbol>()) {
-                    SymbolDisplay.DisplaySymbol(text, typeMember);
-                    text.Write(CreateLine());
-                }
-
-                text.indent--;
-                text.Write(CreatePunctuation(SyntaxKind.CloseBraceToken));
-                text.Write(CreateLine());
-            } catch (BelteInternalException) {
-                if (writeEnding) {
-                    text.Write(CreatePunctuation(SyntaxKind.SemicolonToken));
-                    text.Write(CreateLine());
-                }
-            }
-        }
-
-        if (symbol is MethodSymbol f) {
-            SymbolDisplay.DisplaySymbol(text, f);
-
-            try {
-                var body = LookupMethodFromParents(program, f);
-                text.Write(CreateSpace());
-                DisplayText.DisplayNode(text, body);
-            } catch (BelteInternalException) {
-                // If the body could not be found, it probably means it is a builtin
-                // In that case only showing the signature is what we want
-                text.Write(CreatePunctuation(SyntaxKind.SemicolonToken));
-                text.Write(CreateLine());
-            }
-        } else if (symbol is ITypeSymbolWithMembers t) {
-            SymbolDisplay.DisplaySymbol(text, symbol);
-            WriteTypeMembers(t);
-        } else if (symbol is VariableSymbol v) {
-            SymbolDisplay.DisplaySymbol(text, v);
-
-            if (v.type.typeSymbol is ITypeSymbolWithMembers s && v.type.dimensions == 0)
-                WriteTypeMembers(s);
-            else
-                text.Write(CreateLine());
-        }
-    }
-
     /// <summary>
     /// Emits the program to an assembly.
     /// </summary>
@@ -366,7 +248,11 @@ public sealed class Compilation {
         return null;
     }
 
-    private BoundProgram GetProgram() {
+    /// <summary>
+    /// Gets the previous <see cref="BoundProgram" />, and binds a new one.
+    /// </summary>
+    /// <returns>Newly bound <see cref="BoundProgram" />.</returns>
+    internal BoundProgram GetProgram() {
         var _previous = previous == null ? null : previous.GetProgram();
         return Binder.BindProgram(isScript, _previous, globalScope, _transpilerMode);
     }
